requirements.txt
# ============================================================================
# UPI Conversational Analytics Engine - Dependencies
# ============================================================================

# Core Data Processing
pandas==2.1.4
numpy==1.26.2

# Database & Analytics
duckdb==0.9.2

# Vector Store & Embeddings
langchain==0.1.0
langchain-community==0.0.10
chromadb==0.4.22
faiss-cpu==1.7.4
sentence-transformers==2.2.2

# LLM Integration
groq==0.4.1
openai==1.6.1  # For embeddings (optional)

# Web Framework
streamlit==1.29.0
streamlit-chat==0.1.1

# Visualization
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.18.0

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
requests==2.31.0
tqdm==4.66.1

# Testing
pytest==7.4.3
pytest-cov==4.1.0

# Code Quality
black==23.12.1
flake8==7.0.0
isort==5.13.2

# Logging & Monitoring
loguru==0.7.2

# Optional: For advanced features
# scikit-learn==1.3.2  # For additional ML features
# transformers==4.36.2  # If using custom embeddings
